### 2025-10-27 #9314 @editorialbot
[PRE REVIEW]: LightKrylov: Lightweight implementation of Krylov subspace techniques in modern Fortran
**Submitting author:** @loiseaujc (<a href="http://orcid.org/0000-0002-7244-8416">Jean-Christophe Loiseau</a>)
**Repository:** https://github.com/nekStab/LightKrylov
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.1.0
