### 2025-11-13 08b59b4
Started trimming
