+++ paper/paper.md
@@ -29,18 +29,16 @@ bibliography: paper.bib
 
 # Summary
 
-[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package with minimal dependencies providing a collection of Krylov-based methods for solving linear systems, computing the leading eigenpairs of a linear operator as well as its leading singular triplets.
-When solving linear algebra problems, direct methods tend to have a computational cost scaling as $\mathcal{O}(n^3)$ making them impractical for very large-scale problems.
-Even when $A$ is sparse, care needs to be taken when computing matrix factorizations since they can incur an $\mathcal{O}(n^2)$ storage cost in the worst case by introducing fill-in.
-In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted.
-We refer interested readers to @ipsen-1998 for an introduction to Krylov methods, to @saad-2003 for technical details, and to @frantz-2023 for examples in the field of computational fluid dynamics.
+[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package with minimal dependencies providing a collection of Krylov-based methods for large-scale linear algebra.
+Direct methods tend to have a computational cost scaling as $\mathcal{O}(n^3)$ making them impractical for large-scale problems.
+Even when $A$ is sparse, computing matrix factorizations can incur an $\mathcal{O}(n^2)$ storage cost in the worst case by introducing fill-in.
+In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted, see @ipsen-1998, @saad-2003 and to @frantz-2023.
 
 # Statement of need
 
 ## A collection of Krylov-based algorithms in pure modern Fortran
 
-`LightKrylov` aims to provide Fortran users with familiar SciPy-inspired interfaces to widely used Krylov techniques.
-These include:
+`LightKrylov` provides Fortran users with SciPy-inspired interfaces to widely used Krylov techniques, including:
 
 - **Krylov processes:** Arnoldi factorization, Golub-Kahan bidiagonalization, as well as Lanczos tridiagonalization for Hermitian operators.
 - **Krylov methods:**
@@ -49,19 +47,15 @@ These include:
     - **SVD -** restarted Golub-Kahan bidiagonalization.
 
 Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] in Julia or `PETSc` [@petsc-web-page].
-Integrating multi-language libraries into existing Fortran codes often presents significant challenges related to build systems, dependency management, or potential performance overhead from language interfacing ("two-language problem"). In contrast, `LightKrylov` provides a pure Fortran alternative, compliant with the Fortran 2018 standard, and only requires the Fortran standard library `stdlib` as an external runtime dependency.
-It leverages `stdlib` not only for basic utilities but specifically builds upon `stdlib_linalg` for foundational linear algebra operations, ensuring adherence to community standards and benefiting from its rapid, ongoing development.
+Integrating multi-language libraries into existing Fortran codes often presents significant challenges (e.g. dependency management or potential performance overhead for instance).
+In contrast, `LightKrylov` provides a pure Fortran alternative, compliant with the Fortran 2018 standard, and only requires the Fortran standard library `stdlib` as an external runtime dependency.
 
[…]
