### 2025-11-13 38213b5
Further trimming
