+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package with minimal dependencies providing a collection of Krylov-based methods for large-scale linear algebra.
+[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package providing a collection of Krylov-based methods for large-scale linear algebra.
 Direct methods tend to have a computational cost scaling as $\mathcal{O}(n^3)$ making them impractical for large-scale problems.
 Even when $A$ is sparse, computing matrix factorizations can incur an $\mathcal{O}(n^2)$ storage cost in the worst case by introducing fill-in.
 In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted, see @ipsen-1998, @saad-2003 and to @frantz-2023.
@@ -40,22 +40,20 @@ In contrast, Krylov methods only need a function computing the matrix-vector pro
 
 `LightKrylov` provides Fortran users with SciPy-inspired interfaces to widely used Krylov techniques, including:
 
-- **Krylov processes:** Arnoldi factorization, Golub-Kahan bidiagonalization, as well as Lanczos tridiagonalization for Hermitian operators.
+- **Krylov processes:** Arnoldi factorization, Golub-Kahan bidiagonalization, and Lanczos tridiagonalization.
 - **Krylov methods:**
     - **Linear systems -** Conjugate Gradient (CG), Generalized Minimal Residual method (GMRES) and Flexible GMRES.
-    - **Spectral decomposition -** Arnoldi iteration with Krylov-Schur restart for non-Hermitian operators, Lanczos tridiagonalization with thick restart for Hermitian ones.
-    - **SVD -** restarted Golub-Kahan bidiagonalization.
+    - **Spectral decomposition -** Krylov-Schur method for non-Hermitian operators, Lanczos tridiagonalization for Hermitian ones.
+    - **SVD -** Golub-Kahan bidiagonalization.
 
-Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] in Julia or `PETSc` [@petsc-web-page].
-Integrating multi-language libraries into existing Fortran codes often presents significant challenges (e.g. dependency management or potential performance overhead for instance).
-In contrast, `LightKrylov` provides a pure Fortran alternative, compliant with the Fortran 2018 standard, and only requires the Fortran standard library `stdlib` as an external runtime dependency.
-
-Additionally, using `fypp` [@fypp-webpage] as a build-time dependency significantly reduces code duplication and maintenance effort, ensuring consistency across different data types (`single`, `double`, `extended`, and `quadruple precision`) both for `real` and `complex` numbers while keeping the core algorithmic logic centralized.
-Finally, its build process relies on the Fortran package manager `fpm`, greatly facilitating its installation and integration with the modern Fortran ecosystem.
+Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] or `PETSc` [@petsc-web-page].
+Integrating these into existing Fortran codes can however present challenges (e.g. dependency management or potential performance overhead for instance).
+In contrast, `LightKrylov` provides a pure Fortran alternative, compliant with the Fortran 2018 standard, and requiring only the Fortran standard library `stdlib` as an external dependency.
+Finally, its build process relies on the Fortran package manager `fpm`, facilitating its installation and integration with the modern Fortran ecosystem.
 
 ## A focus on abstract linear operators and abstract vectors
 
[…]
