### 2025-11-13 b41294b
Even more trimming.
