+++ paper/paper.bib
@@ -21,6 +21,7 @@
     journal = {Applied Mechanics Reviews},
     volume = {75},
     number = {3},
+    doi = {10.1115/1.4056808},
 }
 
 @book{saad-2003,
@@ -34,6 +35,7 @@
     isbn = {978-0-89871-534-7},
     keywords = {differential equations,iterative methods,partial-numerical
                 solutions,sparse matrices},
+    doi = {10.1137/1.9780898718003},
 }
 
 @article{ipsen-1998,
@@ -47,6 +49,7 @@
     pages = {889--899},
     publisher = {Taylor \& Francis},
     urldate = {2025-01-20},
+    doi = {10.1080/00029890.1998.12004985},
 }
 
 @article{montoison-2023,
@@ -62,6 +65,7 @@
     urldate = {2025-01-20},
     copyright = {http://creativecommons.org/licenses/by/4.0/},
     langid = {english},
+    doi = {10.21105/joss.05187},
 }
 
  @misc{petsc-web-page,
@@ -97,8 +101,6 @@
     pages = {927-940},
     year = {1988},
     doi = {10.1137/0725052},
-    url = {https://doi.org/10.1137/0725052},
-    eprint = {https://doi.org/10.1137/0725052},
     abstract = { We propose two new conjugate-gradient-type methods for the
                 solution of sparse unsymmetric linear systems. We present a new
                 tridiagonalization process for unsymmetric matrices that is
@@ -126,7 +128,7 @@
     month = {06},
     issn = {1070-6631},
     doi = {10.1063/1.2211705},
-    url = {https://doi.org/10.1063/1.2211705},
+    url = {10.1063/1.2211705},
 }
 
 @misc{fypp-webpage,
