### 2025-11-28 fe00f92
Final trimming
