+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package providing a collection of Krylov-based methods for large-scale linear algebra.
+[`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package providing a collection of Krylov methods for large-scale linear algebra.
 Direct methods tend to have a computational cost scaling as $\mathcal{O}(n^3)$ making them impractical for large-scale problems.
 Even when $A$ is sparse, computing matrix factorizations can incur an $\mathcal{O}(n^2)$ storage cost in the worst case by introducing fill-in.
 In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted, see @ipsen-1998, @saad-2003 and to @frantz-2023.
@@ -45,9 +45,9 @@ In contrast, Krylov methods only need a function computing the matrix-vector pro
 - **SVD -** Golub-Kahan bidiagonalization.
 
 Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] or `PETSc` [@petsc-web-page].
-Integrating these into existing Fortran codes can however present challenges (e.g. dependency management or potential performance overhead for instance).
-In contrast, `LightKrylov` provides a pure Fortran alternative, compliant with the Fortran 2018 standard, and requiring only the Fortran standard library `stdlib` as an external dependency.
-Finally, its build process relies on the Fortran package manager `fpm`, facilitating its installation and integration with the modern Fortran ecosystem.
+Integrating these into existing Fortran codes can however present challenges (e.g. dependency management).
+In contrast, `LightKrylov` is a pure Fortran alternative, compliant with the 2018 standard, and requiring only the Fortran standard library `stdlib` as an dependency.
+Finally, its build process relies on the Fortran package manager `fpm`, facilitating its integration with the modern Fortran ecosystem.
 
 ## A focus on abstract linear operators and abstract vectors
 
@@ -84,33 +84,23 @@ mimicking the signature of the (extended) BLAS-1 subroutine `axpby`.
 Abstract linear operators are defined similarly, with two type-bound procedures required to implement the matrix-vector and transpose matrix-vector product.
 Using `abstract` types enables us to focus on the high-level implementation of the different algorithms while leaving the performance-critical details to the users.
 In addition, `LightKrylov` exposes abstract types for preconditioners, as well as a Newton-GMRES solver for nonlinear systems.
-After extending these `abstract` types for their application, one can solve linear systems or compute eigenvalues as easily as
-
-```fortran
-! Solve linear system.
-call gmres(A, b, x, info)
-
-! Compute eigenvalues and eigenvectors.
-call eigs(A, V, lambda, residuals, info)
-```
-
-<!-- Examples illustrating how to extend these `abstract` types for computing the leading eigenpairs of the linearized Ginzburg-Landau equation or computing unstable periodic orbits of the Rössler system and studying their Lyapunov exponents can be found [here](https://github.com/nekStab/LightKrylov/tree/main/example). -->
+After extending these `abstract` types for their application, one can solve linear systems or compute eigenvalues as easily as `call gmres(A, b, x, info)` or `call eigs(A, V, lambda, residuals, info)`.
 
 # Performance in a production-ready open-source codebase
 
[…]
