### 2025-11-28 #9314 @loiseaujc
@danielskatz : Sorry for the delay. Things have been a bit complicated lately. I've been able to reduce the word count down to 1131 (from 1800) and added the missing doi. Let me know if you need anything else.

Regards
