### 2025-12-09 #9314 @loiseaujc
Thanks @danielskatz. I'm in a PhD committee today. I'll look though the list tomorrow and give you a shortlist as soon as possible. In the mean time, I'm pinging @Simkern (one of my co-authors) so he can keep updated as well.

@Simkern: Can you take a look at the list whenever you have time ? We'll cross-check each of our list on Thursday morning before sending the final one.
