### 2025-12-15 fcca2b9
Check allocations for vectors.
