+++ src/AbstractTypes/AbstractVectors.f90
@@ -929,7 +929,14 @@ contains
     subroutine dense_zero_rsp(self)
         implicit none(type, external)
         class(dense_vector_rsp), intent(inout) :: self
-        if(.not. allocated(self%data)) allocate(self%data(self%n))
+        integer :: iostat
+        character(len=100) :: errmsg
+        if(.not. allocated(self%data)) then
+            allocate(self%data(self%n), stat=iostat, errmsg=errmsg)
+            if (iostat /= 0) call stop_error(errmsg, &
+                                             module=this_module, &
+                                             procedure="dense_zero_rsp")
+        endif
         self%data = zero_rsp
     end subroutine dense_zero_rsp
 
@@ -937,6 +944,8 @@ contains
         implicit none(type, external)
         class(dense_vector_rsp), intent(inout) :: self
         logical, optional, intent(in) :: ifnorm
+        integer :: iostat
+        character(len=100) :: errmsg
         call random_number(self%data)
     end subroutine dense_rand_rsp
 
@@ -954,9 +963,15 @@ contains
         real(sp), intent(in) :: alpha, beta
         class(dense_vector_rsp), intent(inout) :: self
         class(abstract_vector_rsp), intent(in) :: vec
-        integer :: n, m
+        integer :: n, m, iostat
+        character(len=100) :: errmsg
         m = vec%get_size()
-        if(.not. allocated(self%data)) allocate(self%data(m), source=zero_rsp)
+        if(.not. allocated(self%data)) then
+            allocate(self%data(m), source=zero_rsp, stat=iostat, errmsg=errmsg)
+            if (iostat /= 0) call stop_error(errmsg, &
+                                             module=this_module, &
+                                             procedure="dense_axpby_rsp")
+        endif
         n = self%get_size()
         if (m /= n) call stop_error("Inconsistent size between the two vectors.")
 
@@ -1001,7 +1016,14 @@ contains
     subroutine dense_zero_rdp(self)
         implicit none(type, external)
         class(dense_vector_rdp), intent(inout) :: self
-        if(.not. allocated(self%data)) allocate(self%data(self%n))
+        integer :: iostat
+        character(len=100) :: errmsg
+        if(.not. allocated(self%data)) then
+            allocate(self%data(self%n), stat=iostat, errmsg=errmsg)
+            if (iostat /= 0) call stop_error(errmsg, &
+                                             module=this_module, &
+                                             procedure="dense_zero_rdp")
+        endif
         self%data = zero_rdp
     end subroutine dense_zero_rdp
 
@@ -1009,6 +1031,8 @@ contains
         implicit none(type, external)
         class(dense_vector_rdp), intent(inout) :: self
         logical, optional, intent(in) :: ifnorm
+        integer :: iostat
+        character(len=100) :: errmsg
         call random_number(self%data)
     end subroutine dense_rand_rdp
 
@@ -1026,9 +1050,15 @@ contains
         real(dp), intent(in) :: alpha, beta
         class(dense_vector_rdp), intent(inout) :: self
         class(abstract_vector_rdp), intent(in) :: vec
-        integer :: n, m
+        integer :: n, m, iostat
+        character(len=100) :: errmsg
         m = vec%get_size()
-        if(.not. allocated(self%data)) allocate(self%data(m), source=zero_rdp)
+        if(.not. allocated(self%data)) then
+            allocate(self%data(m), source=zero_rdp, stat=iostat, errmsg=errmsg)
+            if (iostat /= 0) call stop_error(errmsg, &
+                                             module=this_module, &
+                                             procedure="dense_axpby_rdp")
+        endif
         n = self%get_size()
         if (m /= n) call stop_error("Inconsistent size between the two vectors.")
 
@@ -1073,7 +1103,14 @@ contains
[…]
