### 2025-12-15 c8be827
Check allocations for linops.
