+++ src/AbstractTypes/AbstractLinops.f90
@@ -902,9 +902,15 @@ contains
 
         ! Working array.
         class(abstract_vector_rsp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="axpby_matvec_rsp")
+        call wrk%zero()
 
         ! w = A @ x
         if (self%transA) then
@@ -932,9 +938,15 @@ contains
 
         ! Working array.
         class(abstract_vector_rsp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="axpby_rmatvec_rsp")
+        call wrk%zero()
 
         ! w = A @ x
         if (self%transA) then
@@ -962,9 +974,15 @@ contains
 
         ! Working array.
         class(abstract_vector_rdp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="axpby_matvec_rdp")
+        call wrk%zero()
 
         ! w = A @ x
         if (self%transA) then
@@ -992,9 +1010,15 @@ contains
 
         ! Working array.
         class(abstract_vector_rdp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="axpby_rmatvec_rdp")
+        call wrk%zero()
 
         ! w = A @ x
         if (self%transA) then
@@ -1022,9 +1046,15 @@ contains
 
         ! Working array.
         class(abstract_vector_csp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="axpby_matvec_csp")
+        call wrk%zero()
 
         ! w = A @ x
         if (self%transA) then
@@ -1052,9 +1082,15 @@ contains
 
         ! Working array.
         class(abstract_vector_csp), allocatable :: wrk
+        integer :: iostat
+        character(len=100) :: errmsg
 
         ! Allocate working array.
-        allocate(wrk, mold=vec_in) ; call wrk%zero()
+        allocate(wrk, mold=vec_in, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
[…]
