### 2025-12-15 fba881a
Check allocations for Krylov methods.
