+++ src/Krylov/utilities.f90
@@ -36,8 +36,12 @@ contains
     end procedure permcols_array_cdp
 
     module procedure invperm
-        integer :: i
-        allocate(inv_perm(size(perm)), source=0)
+        integer :: i, iostat
+        character(len=100) :: errmsg
+        allocate(inv_perm(size(perm)), source=0, stat=iostat, errmsg=errmsg)
+        if (iostat /= 0) call stop_error(errmsg, &
+                                         module=this_module, &
+                                         procedure="invperm")
         inv_perm(perm) = [(i, i=1, size(perm))]
     end procedure invperm
 
