### 2025-12-15 2583377
Check allocations.
