+++ src/AbstractTypes/AbstractVectors.f90
@@ -933,9 +933,7 @@ contains
         character(len=100) :: errmsg
         if(.not. allocated(self%data)) then
             allocate(self%data(self%n), stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_zero_rsp")
+            call check_allocation(iostat, errmsg, this_module, "dense_zero_rsp")
         endif
         self%data = zero_rsp
     end subroutine dense_zero_rsp
@@ -968,9 +966,7 @@ contains
         m = vec%get_size()
         if(.not. allocated(self%data)) then
             allocate(self%data(m), source=zero_rsp, stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_axpby_rsp")
+            call check_allocation(iostat, errmsg, this_module, "dense_axpby_rsp")
         endif
         n = self%get_size()
         if (m /= n) call stop_error("Inconsistent size between the two vectors.")
@@ -1020,9 +1016,7 @@ contains
         character(len=100) :: errmsg
         if(.not. allocated(self%data)) then
             allocate(self%data(self%n), stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_zero_rdp")
+            call check_allocation(iostat, errmsg, this_module, "dense_zero_rdp")
         endif
         self%data = zero_rdp
     end subroutine dense_zero_rdp
@@ -1055,9 +1049,7 @@ contains
         m = vec%get_size()
         if(.not. allocated(self%data)) then
             allocate(self%data(m), source=zero_rdp, stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_axpby_rdp")
+            call check_allocation(iostat, errmsg, this_module, "dense_axpby_rdp")
         endif
         n = self%get_size()
         if (m /= n) call stop_error("Inconsistent size between the two vectors.")
@@ -1107,9 +1099,7 @@ contains
         character(len=100) :: errmsg
         if(.not. allocated(self%data)) then
             allocate(self%data(self%n), stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_zero_csp")
+            call check_allocation(iostat, errmsg, this_module, "dense_zero_csp")
         endif
         self%data = zero_csp
     end subroutine dense_zero_csp
@@ -1122,9 +1112,7 @@ contains
         character(len=100) :: errmsg
         real(sp), allocatable :: y(:, :)
         allocate(y(size(self%data), 2), stat=iostat, errmsg=errmsg)
-        if (iostat /= 0) call stop_error(errmsg, &
-                                         module=this_module, &
-                                         procedure="dense_rand_csp")
+        call check_allocation(iostat, errmsg, this_module, "dense_rand_csp")
         call random_number(y)
         self%data%re = y(:, 1) ; self%data%im = y(:, 2)
     end subroutine dense_rand_csp
@@ -1148,9 +1136,7 @@ contains
         m = vec%get_size()
         if(.not. allocated(self%data)) then
             allocate(self%data(m), source=zero_csp, stat=iostat, errmsg=errmsg)
-            if (iostat /= 0) call stop_error(errmsg, &
-                                             module=this_module, &
-                                             procedure="dense_axpby_csp")
+            call check_allocation(iostat, errmsg, this_module, "dense_axpby_csp")
         endif
         n = self%get_size()
[…]
