### 2025-12-15 364320c
Check allocations.
