+++ src/IterativeSolvers/GMRES/gmres.f90
@@ -123,7 +123,7 @@ contains
 
         ! Miscellaneous.
         character(len=*), parameter :: this_procedure = 'gmres_rsp'
-        integer :: k
+        integer :: k, iostat
         class(abstract_vector_rsp), allocatable :: dx, wrk
         character(len=256) :: msg
 
@@ -147,12 +147,20 @@ contains
         trans = optval(transpose, .false.)
 
         ! Initialize working variables.
-        allocate(wrk, source=b)       ; call wrk%zero()
-        allocate(V(kdim+1), source=b) ; call zero_basis(V)
-        allocate(H(kdim+1, kdim))   ; H = 0.0_sp
-        allocate(e(kdim+1))         ; e = 0.0_sp
-        allocate(c(kdim))           ; c = 0.0_sp
-        allocate(s(kdim))           ; s = 0.0_sp
+        allocate(wrk, source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
+        call wrk%zero()
+        allocate(V(kdim+1), source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
+        call zero_basis(V)
+        allocate(H(kdim+1, kdim), source=zero_rsp, stat=iostat, errmsg=msg) 
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
+        allocate(e(kdim+1), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
+        allocate(c(kdim), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
+        allocate(s(kdim), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, "gmres_rsp")
 
         ! Initialize metadata and & reset matvec counter
         gmres_meta = gmres_sp_metadata() ; gmres_meta%converged = .false.
@@ -175,7 +183,8 @@ contains
             call V(1)%scal(one_rsp/beta)
             c = 0.0_sp ; s = 0.0_sp
             if (gmres_meta%n_outer == 0) then
-               allocate(gmres_meta%res(1)) ; gmres_meta%res(1) = abs(beta)
+               allocate(gmres_meta%res(1), source=abs(beta), stat=iostat, errmsg=msg) 
+               call check_allocation(iostat, msg, this_module, "gmres_rsp")
                write(msg,'(2(A,E11.4))') 'GMRES(k)   init step     : |res|= ', &
                         & abs(beta), ', tol= ', tol
                call log_information(msg, this_module, this_procedure)
@@ -300,7 +309,7 @@ contains
 
         ! Miscellaneous.
         character(len=*), parameter :: this_procedure = 'gmres_rdp'
-        integer :: k
+        integer :: k, iostat
         class(abstract_vector_rdp), allocatable :: dx, wrk
         character(len=256) :: msg
 
@@ -324,12 +333,20 @@ contains
         trans = optval(transpose, .false.)
 
         ! Initialize working variables.
-        allocate(wrk, source=b)       ; call wrk%zero()
-        allocate(V(kdim+1), source=b) ; call zero_basis(V)
-        allocate(H(kdim+1, kdim))   ; H = 0.0_dp
-        allocate(e(kdim+1))         ; e = 0.0_dp
-        allocate(c(kdim))           ; c = 0.0_dp
-        allocate(s(kdim))           ; s = 0.0_dp
+        allocate(wrk, source=b, stat=iostat, errmsg=msg)
[…]
