### 2025-12-15 3e911b1
Check allocations.
