+++ src/IterativeSolvers/GMRES/fgmres.f90
@@ -124,7 +124,7 @@ contains
 
         ! Miscellaneous.
         character(len=*), parameter :: this_procedure = 'fgmres_rsp'
-        integer :: k
+        integer :: k, iostat
         class(abstract_vector_rsp), allocatable :: dx, wrk
         character(len=256) :: msg
 
@@ -148,13 +148,23 @@ contains
         trans = optval(transpose, .false.)
 
         ! Initialize working variables.
-        allocate(wrk, source=b)       ; call wrk%zero()
-        allocate(V(kdim+1), source=b) ; call zero_basis(V)
-        allocate(Z(kdim+1), source=b) ; call zero_basis(Z)
-        allocate(H(kdim+1, kdim))   ; H = 0.0_sp
-        allocate(e(kdim+1))         ; e = 0.0_sp
-        allocate(c(kdim))           ; c = 0.0_sp
-        allocate(s(kdim))           ; s = 0.0_sp
+        allocate(wrk, source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        call wrk%zero()
+        allocate(V(kdim+1), source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        call zero_basis(V)
+        allocate(Z(kdim+1), source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        call zero_basis(Z)
+        allocate(H(kdim+1, kdim), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(e(kdim+1), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(c(kdim), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(s(kdim), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
 
         ! Initialize metadata and & reset matvec counter
         fgmres_meta = fgmres_sp_metadata() ; fgmres_meta%converged = .false.
@@ -177,7 +187,8 @@ contains
             call V(1)%scal(one_rsp/beta)
             c = 0.0_sp ; s = 0.0_sp
             if (fgmres_meta%n_outer == 0) then
-               allocate(fgmres_meta%res(1)) ; fgmres_meta%res(1) = abs(beta)
+               allocate(fgmres_meta%res(1), source=abs(beta), stat=iostat, errmsg=msg)
+               call check_allocation(iostat, msg, this_module, this_procedure)
                write(msg,'(2(A,E11.4))') 'FGMRES(k)   init step     : |res|= ', &
                         & abs(beta), ', tol= ', tol
                call log_information(msg, this_module, this_procedure)
@@ -301,7 +312,7 @@ contains
 
         ! Miscellaneous.
         character(len=*), parameter :: this_procedure = 'fgmres_rdp'
-        integer :: k
+        integer :: k, iostat
         class(abstract_vector_rdp), allocatable :: dx, wrk
         character(len=256) :: msg
 
@@ -325,13 +336,23 @@ contains
         trans = optval(transpose, .false.)
 
         ! Initialize working variables.
-        allocate(wrk, source=b)       ; call wrk%zero()
-        allocate(V(kdim+1), source=b) ; call zero_basis(V)
-        allocate(Z(kdim+1), source=b) ; call zero_basis(Z)
[…]
