### 2025-12-15 5561c3f
Check all allocations.
