+++ src/Expm/ExpmLib.f90
@@ -219,7 +219,7 @@ contains
         real(sp) :: err_est, beta
         ! Optional arguments.
         logical :: transpose
-        integer :: nsteps
+        integer :: nsteps, iostat
         character(len=256) :: msg
     
         ! Deals with optional args.
@@ -230,8 +230,14 @@ contains
         info = 0
 
         ! Allocate arrays.
-        allocate(X(nk+1), source=b) ; allocate(H(nk+1, nk+1))
-        allocate(E(nk+1, nk+1)) ; allocate(Xwrk, source=b)
+        allocate(X(nk+1), source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(H(nk+1, nk+1), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(E(nk+1, nk+1), source=H, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(Xwrk, source=b, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
 
         ! Normalize input vector and initialize Krylov subspace.
         beta = b%norm()
@@ -323,7 +329,7 @@ contains
         real(sp) :: err_est
         ! Optional arguments.
         logical :: transpose
-        integer :: nsteps
+        integer :: nsteps, iostat
         character(len=256) :: msg
 
         ! Determine block size.
@@ -337,12 +343,20 @@ contains
         info = 0
 
         ! Allocate arrays.
-        allocate(R(p, p)) ; allocate(perm(p)) ; allocate(ptrans(p))
-        allocate(X(p*(nk+1)), source=B(1)) ; allocate(H(p*(nk+1), p*(nk+1)))
-        allocate(E(p*(nk+1), p*(nk+1)))
+        allocate(R(p, p), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(perm(p), ptrans(p), source=0, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(X(p*(nk+1)), source=B(1), stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(H(p*(nk+1), p*(nk+1)), E(p*(nk+1), p*(nk+1)), source=zero_rsp, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
 
         ! Scratch arrays.
-        allocate(Xwrk(p), source=B) ; allocate(Cwrk(p), source=B(1))
+        allocate(Xwrk(p), source=B, stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
+        allocate(Cwrk(p), source=B(1), stat=iostat, errmsg=msg)
+        call check_allocation(iostat, msg, this_module, this_procedure)
 
         ! Normalize input matrix and initialize Krylov subspace.
         R = 0.0_sp
@@ -475,7 +489,7 @@ contains
         real(dp) :: err_est, beta
         ! Optional arguments.
         logical :: transpose
-        integer :: nsteps
+        integer :: nsteps, iostat
         character(len=256) :: msg
     
         ! Deals with optional args.
@@ -486,8 +500,14 @@ contains
         info = 0
 
         ! Allocate arrays.
[…]
