### 2025-12-15 7edc236
Remove macOS from CI (temporary)
