+++ .github/workflows/ci.yml
@@ -25,8 +25,8 @@ jobs:
           #   toolchain: {compiler: intel-classic, version: '2021.10', flags: ["/fpp /O3"]}
           # - os: windows-latest
           #   toolchain: {compiler: intel, version: '2024.1', flags: ["/fpp /O3"]}
-          - os: macos-14
-            toolchain: {compiler: gcc, version: 13, flags: ['-cpp -O3 -march=native -mtune=native']}
+          # - os: macos-14
+          #   toolchain: {compiler: gcc, version: 13, flags: ['-cpp -O3 -march=native -mtune=native']}
           # - os: macos-12
           #   toolchain: {compiler: intel-classic, version: '2021.10', flags: ['-fpp -O3 -xhost']}
 
