### 2025-12-15 df7c023
Simplified allocations a bit.
