### 2025-12-15 05f1c5e
Abstract Symmetric/Hermitian Linear Operators
