+++ src/AbstractTypes/AbstractLinops.f90
@@ -45,6 +45,9 @@ module LightKrylov_AbstractLinops
         !! Finalize timers and print complete history_info
     end type abstract_linop
 
+    type, abstract, extends(abstract_linop), public :: abstract_sym_linop
+    end type abstract_sym_linop
+
     !------------------------------------------------------------------------------
     !-----     Definition of an abstract real(sp) operator with kind=sp     -----
     !------------------------------------------------------------------------------
@@ -490,31 +493,113 @@ module LightKrylov_AbstractLinops
     !----------------------------------------------------------------
     !-----     Definition of an abstract symmetric operator     -----
     !----------------------------------------------------------------
-    type, abstract, extends(abstract_linop_rsp), public :: abstract_sym_linop_rsp
+    type, abstract, extends(abstract_sym_linop), public :: abstract_sym_linop_rsp
         !! Abstract representation of an abstract symmetric (real valued) linear operator.
     contains
+        private
+        ! User defined procedures
+        procedure(abstract_sym_matvec_rsp), pass(self), deferred, public :: matvec
+        !! Procedure to compute the matrix-vector product \( \mathbf{y} = \mathbf{Ax} \).
+        ! Wrappers including counter increment
+        procedure, pass(self), public :: apply_matvec => apply_sym_matvec_rsp
+        !! Wrapper for matvec including the counter increment
     end type abstract_sym_linop_rsp
+    abstract interface
+        subroutine abstract_sym_matvec_rsp(self, vec_in, vec_out)
+            !! Interface for the matrix-vector product.
+            use lightkrylov_AbstractVectors
+            import abstract_sym_linop_rsp
+            implicit none(type, external)
+            class(abstract_sym_linop_rsp) , intent(inout)  :: self
+            !! Linear operator \(\mathbf{A}\).
+            class(abstract_vector_rsp), intent(in)  :: vec_in
+            !! Vector to be multiplied by \(\mathbf{A}\).
+            class(abstract_vector_rsp), intent(out) :: vec_out
+            !! Result of the matrix-vector product.
+        end subroutine abstract_sym_matvec_rsp
+    end interface
     !----------------------------------------------------------------
     !-----     Definition of an abstract symmetric operator     -----
     !----------------------------------------------------------------
-    type, abstract, extends(abstract_linop_rdp), public :: abstract_sym_linop_rdp
+    type, abstract, extends(abstract_sym_linop), public :: abstract_sym_linop_rdp
         !! Abstract representation of an abstract symmetric (real valued) linear operator.
     contains
+        private
+        ! User defined procedures
+        procedure(abstract_sym_matvec_rdp), pass(self), deferred, public :: matvec
+        !! Procedure to compute the matrix-vector product \( \mathbf{y} = \mathbf{Ax} \).
+        ! Wrappers including counter increment
+        procedure, pass(self), public :: apply_matvec => apply_sym_matvec_rdp
+        !! Wrapper for matvec including the counter increment
     end type abstract_sym_linop_rdp
+    abstract interface
+        subroutine abstract_sym_matvec_rdp(self, vec_in, vec_out)
+            !! Interface for the matrix-vector product.
+            use lightkrylov_AbstractVectors
+            import abstract_sym_linop_rdp
+            implicit none(type, external)
+            class(abstract_sym_linop_rdp) , intent(inout)  :: self
+            !! Linear operator \(\mathbf{A}\).
+            class(abstract_vector_rdp), intent(in)  :: vec_in
[…]
