### 2025-12-16 0084598
Pre-allocated subspace for kexpm (vector form).
