+++ src/Expm/ExpmLib.f90
@@ -187,7 +187,7 @@ module LightKrylov_ExpmLib
     end interface
 
 contains
-    subroutine kexpm_vec_rsp(c, A, b, tau, tol, info, trans, kdim)
+    subroutine kexpm_vec_rsp(c, A, b, tau, tol, info, trans, X, kdim)
         !! Best approximation of \( \exp(\tau \mathbf{A}) \mathbf{b} \) in the computed Krylov subspace
         implicit none(type, external)
         class(abstract_vector_rsp), intent(out) :: c
@@ -203,6 +203,8 @@ contains
         !! Information flag.
         logical, optional, intent(in) :: trans
         !! Use transpose?
+        class(abstract_vector_rsp), target, optional, intent(inout) :: X(:)
+        !! Krylov subspace.
         integer, optional, intent(in) :: kdim
         !! Maximum size of the Krylov subspace.
 
@@ -211,28 +213,38 @@ contains
         integer, parameter :: kmax = 100
         integer :: k, km, kp, nk
         ! Arnoldi factorization.
-        class(abstract_vector_rsp), allocatable :: X(:)
+        class(abstract_vector_rsp), pointer :: Xwrk(:)
         real(sp), allocatable :: H(:, :)
         ! Normaliztion & temporary arrays.
         real(sp), allocatable :: E(:, :)
-        class(abstract_vector_rsp), allocatable :: Xwrk
+        class(abstract_vector_rsp), allocatable :: wrk
         real(sp) :: err_est, beta
         ! Optional arguments.
         logical :: transpose
         integer :: nsteps, iostat
         character(len=256) :: msg
-    
+
+        ! Ensure either optional kdim or optional X is being used.
+        if (present(X)) then
+            nsteps = size(X) - 1
+            Xwrk(1:nsteps+1) => X
+        else
+            nsteps = optval(kdim, kmax)
+            allocate(Xwrk(nsteps+1), source=b, stat=iostat, errmsg=msg)
+            call check_allocation(iostat, msg, this_module, this_procedure)
+        endif
+
         ! Deals with optional args.
         transpose = optval(trans, .false.)
-        nsteps    = optval(kdim, kmax)
         nk        = nsteps
 
         info = 0
 
         ! Allocate arrays.
-        allocate(X(nk+1), Xwrk, &
+        allocate(wrk, &
                  source=b, stat=iostat, errmsg=msg)
         call check_allocation(iostat, msg, this_module, this_procedure)
+
         allocate(H(nk+1, nk+1), E(nk+1, nk+1), &
                  source=zero_rsp, stat=iostat, errmsg=msg)
         call check_allocation(iostat, msg, this_module, this_procedure)
@@ -245,8 +257,8 @@ contains
             err_est = 0.0_sp
             kp = 1
         else
-            call zero_basis(X)
-            call X(1)%add(b) ; call X(1)%scal(one_rsp / beta)
+            call zero_basis(Xwrk)
+            call Xwrk(1)%add(b) ; call Xwrk(1)%scal(one_rsp / beta)
 
             expm_arnoldi: do k = 1, nk
                 km = k-1 ; kp = k+1
@@ -255,7 +267,7 @@ contains
                 E = 0.0_sp
 
                 ! Compute k-th step Arnoldi factorization.
-                call arnoldi(A, X, H, info, kstart=k, kend=k, transpose=transpose)
+                call arnoldi(A, Xwrk, H, info, kstart=k, kend=k, transpose=transpose)
                 call check_info(info, 'arnoldi', this_module, this_procedure)
 
                 ! Compute approximation.
@@ -269,8 +281,8 @@ contains
                 E(:kp, :kp) = expm(tau*H(:kp, :kp))
 
[…]
