### 2025-12-16 169fd3a
Replaced expm with matrix_exp
