### 2025-12-17 83feeae
add paper todos
