+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 [`LightKrylov`](https://github.com/nekStab/LightKrylov) is a Fortran package providing a collection of Krylov methods for large-scale linear algebra.
 Direct methods tend to have a computational cost scaling as $\mathcal{O}(n^3)$ making them impractical for large-scale problems.
 Even when $A$ is sparse, computing matrix factorizations can incur an $\mathcal{O}(n^2)$ storage cost in the worst case by introducing fill-in.
-In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted, see @ipsen-1998, @saad-2003 and to @frantz-2023.
+In contrast, Krylov methods only need a function computing the matrix-vector product $u \leftarrow Av$ (or $u \leftarrow A^H v$) to iteratively construct a *Krylov subspace* [@krylov-1931] from which the solutions can be extracted, see @ipsen-1998, @saad-2003 and @frantz-2023.
 
 # Statement of need
 
@@ -40,14 +40,14 @@ In contrast, Krylov methods only need a function computing the matrix-vector pro
 
 `LightKrylov` provides Fortran users with SciPy-inspired interfaces to widely used Krylov techniques, including:
 
-- **Linear systems -** Conjugate Gradient (CG), Generalized Minimal Residual method (GMRES) and Flexible GMRES.
+- **Linear systems -** Conjugate Gradient (CG), Generalized Minimal Residual method (GMRES) and Flexible GMRES @citeme.
 - **Spectral decomposition -** Arnoldi method (with Krylov-Schur restart) for non-Hermitian operators, Lanczos tridiagonalization for Hermitian ones.
 - **SVD -** Golub-Kahan bidiagonalization.
 
 Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] or `PETSc` [@petsc-web-page].
 Integrating these into existing Fortran codes can however present challenges (e.g. dependency management).
-In contrast, `LightKrylov` is a pure Fortran alternative, compliant with the 2018 standard, and requiring only the Fortran standard library `stdlib` as an dependency.
-Finally, its build process relies on the Fortran package manager `fpm`, facilitating its integration with the modern Fortran ecosystem.
+In contrast, `LightKrylov` is a pure Fortran alternative, compliant with the 2018 standard, and requiring only the Fortran standard library `stdlib` @citeme as an dependency.
+Finally, its build process relies on the Fortran package manager [`fpm`](link?), facilitating its integration with the modern Fortran ecosystem.
 
 ## A focus on abstract linear operators and abstract vectors
 
@@ -74,7 +74,7 @@ For instance, the `abstract_axpby_rdp` interface reads
 ```fortran
 abstract interface
   subroutine abstract_axpby_rdp(alpha, vec, beta, self)
-    double precision, intent(in)    :: alpha, beta
+    double precision, intent(in)              :: alpha, beta
     class(abstract_vector_rdp), intent(in)    :: vec
     class(abstract_vector_rdp), intent(inout) :: self
   end subroutine
@@ -84,25 +84,25 @@ mimicking the signature of the (extended) BLAS-1 subroutine `axpby`.
 Abstract linear operators are defined similarly, with two type-bound procedures required to implement the matrix-vector and transpose matrix-vector product.
 Using `abstract` types enables us to focus on the high-level implementation of the different algorithms while leaving the performance-critical details to the users.
 In addition, `LightKrylov` exposes abstract types for preconditioners, as well as a Newton-GMRES solver for nonlinear systems.
[…]
