### 2025-12-18 nekStab/LightKrylov#222 @alexfikl
[JOSS] Reproducing example from the paper
The paper has an example with a flow past a cylinder that does not seem to exist in the repository. Could it be added in some fashion? That should make it possible to verify the performance claims from the paper.

xref: https://github.com/openjournals/joss-reviews/issues/9623
