### 2025-12-18 nekStab/LightKrylov#222 @loiseaujc
Indeed, the example is not present in the repo as it requires both `Nek5000` and `neklab` to run. We had a link in a previous draft of the paper to this example. I'll let @Simkern give you the link if you want to rerun it as I can't find it right now (at home with sick kids).

Note that the cylinder flow example actually contains two separate solvers combinations:

- `LightKrylov` + `neklab` + `Nek5000`
- `Nek5000` + `KTHFramework`

They moreover rely on different combinations of build systems (`fpm` + `make` + a `bash` script for the `LightKrylov`/`neklab` one, and only `make` for the `KTHFramework` one). Rather than including both of these directly in the `LightKrylov` repo, we could create a new standalone repo `LightKrylov_examples` where each folder is a self-contained example with its own set of dependencies and build script. In my opinion, this option would keep the `LightKrylov` repo to the point, while also making it easy in the near future to add a large set of specialized examples (i.e. beyond `Nek5000`/`neklab`, or even beyond just CFD in general). Would that work for you?
