+++ paper/tile.sh
@@ -1,1 +1,0 @@
-composite -geometry +450+240 \( timings.png -resize 90% \) LK_newton_eigs.png LK_newton_eigs_timings.png
