### 2025-12-19 f83a30d
Added references
