+++ paper/paper.md
@@ -40,13 +40,13 @@ In contrast, Krylov methods only need a function computing the matrix-vector pro
 
 `LightKrylov` provides Fortran users with SciPy-inspired interfaces to widely used Krylov techniques, including:
 
-- **Linear systems -** Conjugate Gradient (CG), Generalized Minimal Residual method (GMRES) and Flexible GMRES.
+- **Linear systems -** Conjugate Gradient (CG), Generalized Minimal Residual method (GMRES) and Flexible GMRESi [@saad:fgmres:siam].
 - **Spectral decomposition -** Arnoldi method (with Krylov-Schur restart) for non-Hermitian operators, Lanczos tridiagonalization for Hermitian ones.
 - **SVD -** Golub-Kahan bidiagonalization.
 
 Libraries exposing more methods exist in other languages, e.g., `Krylov.jl` [@montoison-2023] or `PETSc` [@petsc-web-page].
 Integrating these into existing Fortran codes can however present challenges (e.g. dependency management).
-In contrast, `LightKrylov` is a pure Fortran alternative, compliant with the 2018 standard, and requiring only the Fortran standard library `stdlib` as an dependency.
+In contrast, `LightKrylov` is a pure Fortran alternative, compliant with the 2018 standard, and requiring only the Fortran standard library [`stdlib`](https://stdlib.fortran-lang.org/), as an dependency.
 Finally, its build process relies on the Fortran package manager `fpm`, facilitating its integration with the modern Fortran ecosystem.
 
 ## A focus on abstract linear operators and abstract vectors
@@ -97,7 +97,7 @@ Discretization of the governing equations leads to systems with approximately 17
 All computations were run in parallel on 12 Intel Core Ultra 7 processors and the numerical settings are identical for both libraries.
 
 The unstable fixed point of the nonlinear Navier-Stokes equations is computed using both `LightKrylov`'s *time-stepper*-based Newton-GMRES solver and the selected frequency damping implementation from `KTH Framework` .
-Likewise, the leading eigenpair of the corresponding linearized Navier-Stokes operator is computed using `LightKrylov`'s implementation of the Krylov-Schur algorithm and the `KTH Framework`'s wrapper for `ARPACK`.
+Likewise, the leading eigenpair of the corresponding linearized Navier-Stokes operator is computed using `LightKrylov`'s implementation of the Krylov-Schur algorithm and the `KTH Framework`'s wrapper for `ARPACK` [@lehoucq:arpack:siam].
 
 A visual comparison is provided in \autoref{fig:timings} showing excellent agreement.
 The table in the lower-right panel of \autoref{fig:timings} summarizes the wall-clock times of the `neklab` computations. Isolating the intrinsic cost of the algorithms in `LightKrylov` from the cost of the calls to LAPACK and the linear and nonlinear Navier-Stokes solvers (`matvec` and `response`, respectively) shows that extended `abstract` types and object-oriented programming in Fortran incurs a negligible computational overhead for such large-scale applications.
