+++ paper/paper.bib
@@ -137,3 +137,25 @@
     url = {https://github.com/aradi/fypp},
     year = {2023},
 }
+
+
+@article{saad:fgmres:siam,
+    author = {Saad, Youcef},
+    title = {A Flexible Inner-Outer Preconditioned GMRES Algorithm},
+    journal = {SIAM Journal on Scientific Computing},
+    volume = {14},
+    number = {2},
+    pages = {461-469},
+    year = {1993},
+    doi = {10.1137/0914028},
+}
+
+@inbook{lehoucq:arpack:siam,
+    author = {R. B. Lehoucq, and D. C. Sorensen, and C. Yang},
+    title = {Introduction to ARPACK},
+    booktitle = {ARPACK Users' Guide},
+    pages = {1-7},
+    doi = {10.1137/1.9780898719628.ch1},
+    URL = {https://epubs.siam.org/doi/abs/10.1137/1.9780898719628.ch1},
+    eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9780898719628.ch1},
+}
