### 2025-12-19 e271a71
Added missing date in arpack ref.
