+++ paper/paper.bib
@@ -158,4 +158,5 @@
     doi = {10.1137/1.9780898719628.ch1},
     URL = {https://epubs.siam.org/doi/abs/10.1137/1.9780898719628.ch1},
     eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9780898719628.ch1},
+    year = {1998},
 }
