### 2025-12-20 nekStab/LightKrylov#226 @ewu63
[JOSS] Update documentation links in README
I noticed that the links to the examples are simply pointing to the README page itself. Also, an API documentation site seems to be available [here](https://nekstab.github.io/LightKrylov/) but is only linked to via one badge, it would probably be best to have it as an explicit link somewhere.

As part of openjournals/joss-reviews#9623
