### 2025-12-20 nekStab/LightKrylov#227 @ewu63
[JOSS] Clarify some HPC usages
In the paper, it is stated
> The abstract vector interface allows direct use of  Nek5000’s distributed data structures, and the pure-Fortran nature facilitated integration with its existing build system, demonstrating the library’s suitability for large-scale HPC applications.

Further down it is mentioned
> All computations were run in parallel on 12 Intel Core Ultra 7 processors and the numerical settings are identical for both libraries.

Could you clarify this a little bit? Is there support for MPI-based parallelization? How exactly does this code interact with (distributed-memory?) parallelism, both in terms of the data structures and the algorithms employed.

As part of openjournals/joss-reviews#9623
