### 2026-01-05 nekStab/LightKrylov#228 @joewallwork
[JOSS review] Contribution guidelines
Towards https://github.com/openjournals/joss-reviews/issues/9623

The JOSS review checklist includes
> Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I see that the README includes details of the development team and the statement "Anyone else interested in contributing is obviously most welcomed!", but it'd be good to provide more details on how to do this.
