### 2026-01-05 nekStab/LightKrylov#229 @joewallwork
[JOSS review] Target audience
Towards https://github.com/openjournals/joss-reviews/issues/9623

As with the paper review, it'd be nice if you could include a few words in the README on who the target audience are.
