### 2026-01-05 nekStab/LightKrylov#222 @alexfikl
> On my laptop, the performances of call spmv(A, x, y) and call linop%(vec_in, vec_out) are essentially identical (less than 0.5% difference in computational time).

For reference, I'm getting (compiled with `fpm build --profile release`):
```
 Dimension :       90449 x       90449
 nnz(A)    :     2455670

Meth.: Fortran stdlib
 Elapsed time :     0.008027000 [s]

Meth.: LightKrylov
 Elapsed time :     0.008375100 [s]
```
