### 2026-01-12 9339b46
Updated examples links (Ginzburg-Landau and Rossler)
