### 2026-01-12 113f78c
Added templates for the different contributing pages.
