### 2026-01-12 655cd92
added visualisations for the GL and Roessler examples.
