### 2026-01-12 dc88a37
Updated from stdlib to lightkrylov links.
