+++ CONTRIBUTING.md
@@ -1,15 +1,15 @@
-# Contributing to the Fortran standard library
+# Contributing to the LightKrylov
 
-Thank you for considering contributing to the Fortran standard library (*stdlib*).
+Thank you for considering contributing to LightKrylov.
 Please review and follow these guidelines to make the contribution process
 simple and effective for all involved. It will help communicate that you
 respect the time of the community developers. In return, the community will
 help to address your problem, evaluate changes, and guide you through your pull
 requests.
 
-By contributing to *stdlib*, you certify that you own or are allowed to share the
+By contributing to *LightKrylov*, you certify that you own or are allowed to share the
 content of your contribution under the
-[stdlib license](https://github.com/fortran-lang/stdlib/blob/HEAD/LICENSE).
+[LightKrylov license](https://github.com/nekStab/LightKrylov/blob/update_readme/LICENSE).
 
 * [Style](#style)
 * [Reporting a bug](#reporting-a-bug)
@@ -21,15 +21,15 @@ content of your contribution under the
 
 ## Style
 
-Please follow the
+The development of `LightKrylov` follows the same style guide as the [`Fortran standard library`](https://github.com/fortran-lang/stdlib). Please follow the
 [Fortran stdlib style guide](https://github.com/fortran-lang/stdlib/blob/HEAD/STYLE_GUIDE.md)
 for any Fortran code that you contribute.
 This allows the community to focus on substance rather than style.
 
 The style guide is a living document.
 You are welcome to propose changes to the style guide by
-[opening an issue](https://github.com/fortran-lang/stdlib/issues/new/choose) or
-[starting a discussion](https://github.com/fortran-lang/stdlib/discussions/new).
+[opening an issue](https://github.com/nekStab/LightKrylov/issues/new/choose) or
+[starting a discussion](https://github.com/nekStab/LightKrylov/discussions/new).
 
 
 ## Reporting a bug
@@ -48,7 +48,7 @@ Before opening a bug report:
 A good bug report should include all information needed to reproduce the bug.
 Please be as detailed as possible:
 
-1. Which version of *stdlib* are you using?
+1. Which version of *LightKrylov* are you using?
    Which compiler version are you using?
    Which platform and architecture are you on?
    Please be specific.
@@ -70,16 +70,15 @@ mocked-up snippet of what the output or behavior would look like with this
 feature implemented. “Crazy,” out-of-the-box ideas are especially welcome.
 It is quite possible we have not considered such solutions yet.
 
-
 ## Workflow
 
 The general workflow is documented in
-[this document](https://github.com/fortran-lang/stdlib/blob/HEAD/WORKFLOW.md)
+[this document](https://github.com/nekStab/LightKrylov/blob/HEAD/WORKFLOW.md)
 
 The workflow guide is a living document.
 You are welcome to propose changes to the workflow by
-[opening an issue](https://github.com/fortran-lang/stdlib/issues/new/choose) or
-[starting a discussion](https://github.com/fortran-lang/stdlib/discussions/new).
+[opening an issue](https://github.com/nekStab/LightKrylov/issues/new/choose) or
+[starting a discussion](https://github.com/nekStab/LightKrylov/discussions/new).
 
 
 ## General guidelines
@@ -90,18 +89,7 @@ You are welcome to propose changes to the workflow by
   merge cycle.
 * Add tests for your feature or bug fix to be sure that it stays functional and useful.
 * Include new features and changes in the
-  [CHANGELOG](https://github.com/fortran-lang/stdlib/blob/master/CHANGELOG.md)
+  [CHANGELOG](https://github.com/nekStab/LightKrylov/blob/master/CHANGELOG.md)
[…]
