+++ WORKFLOW.md
@@ -2,17 +2,13 @@
 
 This document describes our current workflow.
 
-We welcome everyone and anyone to participate and propose additions to stdlib.
+We welcome everyone and anyone to participate and propose additions to LightKrylov.
 It is okay if you do not have experience for specification or implementation,
-but have an idea for stdlib. If the idea is popular among the community, more
+but have an idea for LightKrylov. If the idea is worth pursuing, more
 experienced contributors will help it through all 5 steps.
 
 
-1. **Idea**: You have an idea or a proposal. Open an
-   [issue](https://github.com/fortran-lang/stdlib/issues) to discuss it. This
-   is on the level of "is there interest in having image reader/writer
-   functions in stdlib?" The goal of this step is to find out if the community
-   is interested in having this functionality as part of stdlib.
+1. **Idea**: You have an idea or a proposal. Open an [issue](https://github.com/nekStab/LightKrylov/issues) to discuss it. This is on the level of "is there interest in having this procedure in LightKrylov?" The goal of this step is to find out if the community is interested in having this functionality as part of LightKrylov.
 
 2. **API**: When there seems to be significant interest in the proposal (vast
    majority of participants think it is a good idea), move on to discuss the
@@ -32,7 +28,7 @@ experienced contributors will help it through all 5 steps.
    [specification is a document](https://stdlib.fortran-lang.org/page/specs/index.html)
    that describes the API and
    the functionality, so that anyone can use it to create an implementation
-   from scratch without looking at `stdlib`. The `stdlib` library then provides
+   from scratch without looking at `LightKrylov`. The `LightKrylov` library then provides
    the reference implementation.
 
 4. **Implementation** in experimental: When opening a PR, request reviews from
@@ -47,10 +43,7 @@ experienced contributors will help it through all 5 steps.
 5. **Release**: Moving from experimental to release. The experimental
    "namespace" contains new functionality together with its specification. In
    order to move from experimental to release, the specification document must
-   be approved by the wide community and the standards committee (informally).
-   If that happens, it has now been blessed for broad use and we can move the
-   code into the main section of `stdlib`, and the particular specification
-   document becomes part of the Fortran Standard Library.
+   be approved by the wide community.
 
 
 Note: the general term "vast majority" above means at least 80%, but ultimately
@@ -58,95 +51,9 @@ it is left to our best judgement to ensure that the community agrees that each
 PR and proposal was approved by "vast majority".
 
 You are welcome to propose changes to this workflow by opening an
-[issue](https://github.com/fortran-lang/stdlib/issues).
+[issue](https://github.com/nekStab/LightKrylov/issues).
 
 
 ## Build systems
 
-This project supports two build systems,
-[fpm](https://github.com/fortran-lang/fpm) and CMake.
-
-### CMake build files
-
-The build files for CMake allow both in-tree, *i.e.* build artifacts share
-the same tree as the source files, and out-of-tree builds, *i.e.* build artifacts
-exist in a separate directory tree.
-Both build types are explicitly supported and tested, the latter strategy
-is recommended for local development.
-
-Sources for the main library target are added in ``src/CMakeLists.txt``
-relative to the library target, *i.e.* no absolute paths are required.
-
-To add tests, the macro ``ADDTEST`` should be used instead of the CMake function
-``add_test``, the macro hides creation of the executable target, linking against the 
-main library target and registering the test.
-The tests themselves are defined as standalone executables in the subdirectories
-in ``test``, a new subdirectory with tests has to be registered in
-``test/CMakeLists.txt``.
-
[…]
