### 2026-01-12 0863875
Added a scope statement.
