### 2026-01-12 9bedb7c
Update CONTRIBUTING.md
minute changes in wording.
