### 2026-01-12 dadf71d
Update WORKFLOW.md
small wording changes
