+++ CODE_OF_CONDUCT.md
@@ -55,8 +55,8 @@ be further defined and clarified by project maintainers.
 ## Enforcement
 
 Instances of abusive, harassing or otherwise unacceptable behavior may be
-reported by contacting one of the project maintainers at caomaco@gmail.com or 
-ondrej@certik.us. All complaints will be reviewed and investigated and will
+reported by contacting one of the project maintainers at jean-christophe.loiseau@ensam.eu or 
+johann-simon.kern@ensam.eu. All complaints will be reviewed and investigated and will
 result in a response that is deemed necessary and appropriate to the
 circumstances. The project team is obligated to maintain confidentiality with
 regard to the reporter of an incident. Further details of specific enforcement
