### 2026-01-19 67dfd22
Reworked summary paragraph.
