### 2026-01-19 b178f30
Added paragraph comparing LK with PETSc and Krylov.jl
