### 2026-01-19 596ea28
Updated PETSc reference
