+++ paper/paper.bib
@@ -68,16 +68,23 @@
     doi = {10.21105/joss.05187},
 }
 
- @misc{petsc-web-page,
-    author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Jed
-              Brown and Peter Brune and Kris Buschelman and Lisandro Dalcin and
-              Victor Eijkhout and William~D. Gropp and Dinesh Kaushik and Matthew
-              ~G. Knepley and Lois Curfman McInnes and Karl Rupp and Barry~F.
-              Smith and Stefano Zampini and Hong Zhang},
-    title = {{PETSc} web page},
-    url = {http://www.mcs.anl.gov/petsc},
-    year = {2015},
+@misc{petsc-web-page,
+    author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven
+              Benson and Jed Brown and Peter Brune and Kris Buschelman and Emil~
+              M. Constantinescu and Lisandro Dalcin and Alp Dener and Victor
+              Eijkhout and Jacob Faibussowitsch and William~D. Gropp and V\'{a}
+              clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev
+              and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott
+              Kruger and Dave~A. May and Lois Curfman McInnes and Richard Tran
+              Mills and Lawrence Mitchell and Todd Munson and Jose~E. Roman and
+              Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith and
+              Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
+    title = {{PETS}c {W}eb page},
+    url = {https://petsc.org/},
+    howpublished = {\url{https://petsc.org/}},
+    year = {2025},
 }   
+
 @article{kth-framework,
     title = {A comprehensive framework to enhance numerical simulations in the
              spectral-element code Nek5000},
