+++ project-file.md
@@ -23,11 +23,9 @@ css: ./bootstrap.min.css
 ---
 
 @warning
-   This API documentation for the `LightKrylov` pacakge is a work in progress.
-   It is build from the source code in the `main` branch and does not track the current development in `dev` or any other branches.
+   This API documentation for the `LightKrylov` package is build from the source code in the `main` branch and does not track the current development in `dev` or any other branches.
    If you use another branch, please refer to the in-code documentation.
    Use the navigation bar at the top of the screen to browse `modules`, `procedures`, `source files`, etc.
-   The listings near the bootom of the page are incomplete.
 @endwarning
 
 This is the main API documentation landing page generated by [FORD](https://github.com/Fortran-FOSS-Programmers/ford#readme). This documentation is released under the [`CC-BY-SA`](https://creativecommons.org/licenses/by-sa/4.0/) license while the `LightKrylov` source code is distribution under the [`BSD-3 Clause`](https://opensource.org/license/bsd-3-clause) one.
