### 2026-01-19 6daef3c
Update paper.md
typos
