### 2026-01-19 8c7f358
Update project-file.md
typos
