### 2026-01-20 1cbc296
Removed ifx as not supported by github action setup-fortran.
