+++ .github/workflows/ci.yml
@@ -8,21 +8,25 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-14]
+        os: [ubuntu-latest, macos-latest]
         toolchain:
           - {compiler: gcc, version: 10}
           - {compiler: gcc, version: 11}
           - {compiler: gcc, version: 12}
           - {compiler: gcc, version: 13}
           - {compiler: gcc, version: 14}
-          - {compiler: intel, version: '2024.1'}
-        include:
-          - os: ubuntu-22.04
-            toolchain: {compiler: intel-classic, version: '2021.10'}
+          - {compiler: gcc, version: 15}
+          - {compiler: intel, version: '2025.0'}
+          - {compiler: intel-classic, version: '2021.10'}
+          - {compiler: nvidia-hpc, version: '25.1'}
         exclude:
-          - os: macos-14
-            toolchain: {compiler: intel, version: '2024.1'}
-          - os: macos-14
+          - os: macos-latest
+            toolchain: {compiler: intel}
+          - os: macos-latest
+            toolchain: {compiler: intel-classic}
+          - os: macos-latest
+            toolchain: {compiler: nvidia-hpc}
+          - os: macos-latest
             toolchain: {compiler: gcc, version: 10}
 
     steps:
