### 2026-01-20 17179b3
Fixing gcc 10 compilation error.
