+++ .github/workflows/ci.yml
@@ -15,10 +15,13 @@ jobs:
           - {compiler: gcc, version: 12}
           - {compiler: gcc, version: 13}
           - {compiler: gcc, version: 14}
+          - {compiler: gcc, version: 15}
           - {compiler: intel, version: '2025.0'}
           - {compiler: intel-classic, version: '2021.10'}
           - {compiler: nvidia-hpc, version: '25.1'}
         exclude:
+          - os: ubuntu-latest
+            toolchain: {compiler: gcc, version: 15}
           - os: macos-latest
             toolchain: {compiler: intel}
           - os: macos-latest
