### 2026-01-20 6024771
Added windows env to the CI.
