+++ .github/workflows/ci.yml
@@ -8,7 +8,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         toolchain:
           - {compiler: gcc, version: 10}
           - {compiler: gcc, version: 11}
@@ -30,6 +30,8 @@ jobs:
             toolchain: {compiler: nvidia-hpc}
           - os: macos-latest
             toolchain: {compiler: gcc, version: 10}
+          - os: windows-latest
+            toolchain: {compiler: nvidia-hpc}
 
     steps:
       - name: Checkout code
