+++ src/AbstractVectors.f90
@@ -1,0 +1,1911 @@
+module LightKrylov_AbstractVectors
+    !! This module provides the base class `absract_vector` from which all Krylov vectors
+    !! needs to be derived. To use `LightKrylov`, you need to extend one of the
+    !! followings:
+    !!
+    !! - `abstract_vector_rsp`  :   Real-valued vector with single precision arithmetic.
+    !! - `abstract_vector_rdp`  :   Real-valued vector with double precision arithmetic.
+    !! - `abstract_vector_csp`  :   Complex-valued vector with single precision arithmetic.
+    !! - `abstract_vector_cdp`  :   Complex-valued vector with double precision arithmetic.
+    !!
+    !! To extend either of these abstract types, you need to provide an associated implementation
+    !! for the following type-bound procedures:
+    !!
+    !! - `zero(self)`                   :   A subroutine zeroing-out the vector.
+    !! - `rand(self, ifnorm)`           :   A subroutine creating a random vector, possibily normalized to have unit-norm (`ifnorm = .true.`).
+    !! - `scal(self, alpha)`            :   A subroutine computing *in-place* the scalar multiplication \( \mathbf{x} \leftarrow \alpha \mathbf{x} \).
+    !! - `axpby(alpha, vec, beta, self) :   A subroutine computing *in-place* the product \( \mathbf{y} \leftarrow \alpha \mathbf{x} + \beta \mathbf{y} \).
+    !! - `dot(self, vec)`               :   A function computing the inner product \( \alpha = \langle \mathbf{x} \vert \mathbf{y} \rangle \).
+    !! - `get_size(self)`               :   A function returning the dimension \( n \) of the vector \( \mathbf{x} \).
+    !!
+    !! Once these type-bound procedures have been implemented by the user, they will automatically 
+    !! be used to define:
+    !!
+    !! - vector addition    :   `add(self, vec) = axpby(1, vec, 1, self)`
+    !! - vector subtraction :   `sub(self, vec) = axpby(-1, vec, 1, self)`
+    !! - vector norm        :   `norm(self)     = sqrt(self%dot(self))`
+    !!
+    !! This module also provides the following utility subroutines:
+    !!
+    !! - `innerprod(X, Y)`                  : Function computing the product \(\mathbf{X}^H \mathbf{y} \) between a Krylov basis `X` and a Krylov vector  (resp. basis) `Y`.
+    !! - `linear_combination(Y, X, V)`      : Subroutine computing the linear combination \( \mathbf{y}_j = \sum_{i=1}^n \mathbf{x}_i v_{ij} \).
+    !! - `axpby_basis(alpha, X, beta, Y)`   : In-place computation of \( \mathbf{Y} \leftarrow \alpha \mathbf{X} + \beta \mathbf{Y} \) where `X` and `Y` are arrays of `abstract_vector`.
+    !! - `zero_basis(X)`                    : Zero-out a collection of `abstract_vectors`.
+    !! - `copy_basis(out, from)`            : Copy a collection of `abstract_vectors`.
+    !! - `rand_basis(X, ifnorm)`            : Create a collection of random `abstract_vectors`. If `ifnorm = .true.`, the vectors are normalized to have unit-norm.
+
+    use stdlib_optval, only: optval
+    use stdlib_linalg_blas, only: scal, axpy, dot, dotc
+    use LightKrylov_Constants
+    use LightKrylov_Utils
+    use LightKrylov_Logger
+    implicit none(type, external)
+    private
+
+    character(len=*), parameter :: this_module      = 'LK_Vectors'
+    character(len=*), parameter :: this_module_long = 'Lightkrylov_AbstractVectors'
+
+    public :: innerprod, Gram
[…]
