+++ src/utilities.f90
@@ -1,0 +1,207 @@
+submodule (lightkrylov_basekrylov) krylov_utilities
+    implicit none(type, external)
+contains
+
+    !----------------------------------------
+    !-----     Permutation matrices     -----
+    !----------------------------------------
+
+    module procedure permcols_basis_rsp
+        call copy(Q, Q(perm))
+    end procedure permcols_basis_rsp
+
+    module procedure permcols_array_rsp
+        Q = Q(:, perm)
+    end procedure permcols_array_rsp
+    module procedure permcols_basis_rdp
+        call copy(Q, Q(perm))
+    end procedure permcols_basis_rdp
+
+    module procedure permcols_array_rdp
+        Q = Q(:, perm)
+    end procedure permcols_array_rdp
+    module procedure permcols_basis_csp
+        call copy(Q, Q(perm))
+    end procedure permcols_basis_csp
+
+    module procedure permcols_array_csp
+        Q = Q(:, perm)
+    end procedure permcols_array_csp
+    module procedure permcols_basis_cdp
+        call copy(Q, Q(perm))
+    end procedure permcols_basis_cdp
+
+    module procedure permcols_array_cdp
+        Q = Q(:, perm)
+    end procedure permcols_array_cdp
+
+    module procedure invperm
+        integer :: i, iostat
+        character(len=100) :: errmsg
+        allocate(inv_perm(size(perm)), source=0, stat=iostat, errmsg=errmsg)
+        call check_allocation(iostat, errmsg, this_module, "invperm")
+        inv_perm(perm) = [(i, i=1, size(perm))]
+    end procedure invperm
+
+    !----------------------------------------------
+    !-----     Initialize Krylov subspace     -----
+    !----------------------------------------------
+
+    module procedure initialize_krylov_subspace_rsp
+        integer :: p
+
+        ! Zero-out X.
+        call zero_basis(X)
+
+        ! Deals with optional args.
+        if(present(X0)) then
+            p = size(X0)
+            ! Initialize.
+            call copy(X(:p), X0)
+            ! Orthonormalize.
+            call orthonormalize_basis(X(:p))
+        endif
+    end procedure initialize_krylov_subspace_rsp
+    module procedure initialize_krylov_subspace_rdp
+        integer :: p
+
+        ! Zero-out X.
+        call zero_basis(X)
+
+        ! Deals with optional args.
+        if(present(X0)) then
+            p = size(X0)
+            ! Initialize.
+            call copy(X(:p), X0)
+            ! Orthonormalize.
+            call orthonormalize_basis(X(:p))
+        endif
+    end procedure initialize_krylov_subspace_rdp
+    module procedure initialize_krylov_subspace_csp
+        integer :: p
+
+        ! Zero-out X.
+        call zero_basis(X)
+
+        ! Deals with optional args.
+        if(present(X0)) then
+            p = size(X0)
+            ! Initialize.
+            call copy(X(:p), X0)
+            ! Orthonormalize.
+            call orthonormalize_basis(X(:p))
+        endif
+    end procedure initialize_krylov_subspace_csp
+    module procedure initialize_krylov_subspace_cdp
+        integer :: p
+
+        ! Zero-out X.
+        call zero_basis(X)
+
+        ! Deals with optional args.
+        if(present(X0)) then
+            p = size(X0)
+            ! Initialize.
+            call copy(X(:p), X0)
+            ! Orthonormalize.
+            call orthonormalize_basis(X(:p))
+        endif
+    end procedure initialize_krylov_subspace_cdp
+   
+    !----------------------------------------
+    !-----     Orthonormalize basis     -----
+    !----------------------------------------
+
[…]

[…]
