### 2026-01-20 34c94eb
Updated the list of tested compiler/OS combinations.
