### 2026-01-20 nekStab/LightKrylov#229 @joewallwork
Yes, I'm happy for this issue to be closed by the PR. Maybe the following integration works?

Closed by #234.
