### 2026-01-20 #9623 @danielskatz
@loiseaujc - I see that you've updated the paper. Do you have any other things you are working on, or have you addressed all the reviewer points now?
