### 2026-01-22 #9623 @loiseaujc
@danielskatz - I think we have addressed everything so far. I'll give a last polishing round to the paper today (fixing a couple of typos and adding a reference to the Fortran standard library) and I think we're done. Really good experience so far !
