+++ paper/paper.bib
@@ -167,3 +167,21 @@
     eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9780898719628.ch1},
     year = {1998},
 }
+
+@article{stdlib:ieee,
+    author = {Kedward, Laurence J. and Aradi, Bálint and Čertík, Ondřej and
+              Curcic, Milan and Ehlert, Sebastian and Engel, Philipp and Goswami,
+              Rohit and Hirsch, Michael and Lozada-Blanco, Asdrubal and Magnin,
+              Vincent and Markus, Arjen and Pagone, Emanuele and Pribec, Ivan and
+              Richardson, Brad and Snyder, Harris and Urban, John and Vandenplas,
+              Jérémie},
+    journal = {Computing in Science & Engineering},
+    title = {The State of Fortran},
+    year = {2022},
+    volume = {24},
+    number = {2},
+    pages = {63-72},
+    keywords = {Standards;Codes;Computer
+                languages;Syntactics;Arrays;Ecosystems;Parallel processing},
+    doi = {10.1109/MCSE.2022.3159862},
+}
