### 2026-01-22 #9623 @joewallwork
> [@editorialbot](https://github.com/editorialbot) generate pdf
>
> [@danielskatz](https://github.com/danielskatz), [@joewallwork](https://github.com/joewallwork), [@alexfikl](https://github.com/alexfikl), [@ewu63](https://github.com/ewu63) : Here is the latest version of the paper. An academic reference to the Fortran standard library has been added, links to the github profiles of the various stdlib members we acknowledge, and we homogenised the -zed vs -sed (US vs UK) stuff.
>
> [@joewallwork](https://github.com/joewallwork) and [@ewu63](https://github.com/ewu63) : _Missing **reproducibility** and **performances** checkmarks in your checklist_
>
>     * We've added the [spmv](https://github.com/nekStab/LightKrylov_examples/tree/main/spmv) to illustrate the negligible overhead of object-oriented programming and abstract types in Fortran using a sparse matrix-vector product example. The matrix has dimension 90 000-ish and comes from Matrix Market. The `spmv` kernel is directly taken from `stdlib_sparse`.
>
>     * The [`neklab`](https://github.com/nekStab/LightKrylov_examples/tree/main/neklab) example contains all of the scripts to reproduce the fluid dynamics computation presented in the paper. As far as I know, I think [@alexfikl](https://github.com/alexfikl) has been able to re-run it with no problem. Let us know if you encounter any issue and we'll try to streamline the example. Note that the build system from `neklab` still is pretty debatable and is essentially hijacking that of `Nek5000`. We plan to improve this at some point.
>
>     * We also added the [`poisson`](https://github.com/nekStab/LightKrylov_examples/tree/main/poisson) illustrating how to implement a simple Poisson solver on the unit square (with homogeneous Dirichlet boundary condition) using a classical MPI-based domain decomposition and `LightKrylov`'s preconditioned CG solver.

Great, thanks! I was planning to look into this tomorrow so this info is helpful.
