### 2026-01-22 #9623 @alexfikl
> * As far as I know, I think [@alexfikl](https://github.com/alexfikl) has been able to re-run it with no problem.

@loiseaujc Just to clarify, I managed to run the `spmv` example very easily. I just quickly tried the `neklab` example and haven't managed to get it to work (not quite sure what order to run the various scripts in). It would be very helpful to add some "install" instruction to the `README.md` file in there :grin:

EDIT: Actually, just move the `examples/README.md`? I'm trying the instructions from there now and will report back.
