### 2026-01-22 #9623 @loiseaujc
Ah ok, I thought you ran the example. My bad.
As for the README for `neklab`, I am working on it as we speak. I hope to have it wrapped up by tonight.
