### 2026-01-22 #9623 @alexfikl
> As for the README for `neklab`, I am working on it as we speak. I hope to have it wrapped up by tonight.

Sounds good! I opened a little issue in `LightKrylov_examples` with the point I got stuck at right now and will try again tomorrow (or whenever the README is updated).
