### 2026-01-22 #9623 @loiseaujc
I tagged @SimKern on the `LightKrylov_example` issue. He might be able to answer quite quickly. Gotta go lecturing.
