### 2026-01-22 #9623 @alexfikl
> I tagged [@Simkern](https://github.com/Simkern) on the `LightKrylov_example` issue. He might be able to answer quite quickly. Gotta go lecturing.

With the latest updates in the `LightKrylov_examples` repository, I managed to compile and run the example from the paper as well. Thanks for the quick fixes!

With that in mind, I gladly re-iterate my recommendation for publication. :grin:
