### 2026-01-23 #9623 @joewallwork
I had a go at setting up the examples to do further testing but hit a compilation error, which I raised an issue for.

Regarding the paper, it looks good to me and I only have minor comments:
* I think "Youcef" Saad is a typo in the 1993 reference.
* julia->Julia and krylov->Krylov in Montoison 2023 reference.
