+++ paper/paper.bib
@@ -53,7 +53,7 @@
 }
 
 @article{montoison-2023,
-    title = {Krylov.jl: A Julia basket of hand-picked Krylov methods},
+    title = {{Krylov.jl: A Julia basket of hand-picked Krylov methods}},
     shorttitle = {Krylov.jl},
     author = {Montoison, Alexis and Orban, Dominique},
     year = {2023},
@@ -147,7 +147,7 @@
 
 
 @article{saad:fgmres:siam,
-    author = {Saad, Youcef},
+    author = {Saad, Yousef},
     title = {A Flexible Inner-Outer Preconditioned GMRES Algorithm},
     journal = {SIAM Journal on Scientific Computing},
     volume = {14},
