### 2026-01-26 #9623 @joewallwork
Thanks to @loiseaujc @alexfikl and @Simkern's assistance on issue https://github.com/nekStab/LightKrylov_examples/issues/9, I've now successfully built the examples. I was able to run the Newton-GMRES test case and validate the performance claims in the paper.

I've now completed my checklist and am happy to recommend the submission for publication, subject to my minor requests in the comment above and the closing of issues https://github.com/nekStab/LightKrylov_examples/issues/9 and https://github.com/nekStab/LightKrylov_examples/issues/8.
