### 2026-01-26 #9623 @danielskatz
@ewu63 - I notice you have one item in your review unchecked, perhaps related to https://github.com/nekStab/LightKrylov/issues/238 - how do you think this can be resolved?
