### 2026-01-26 #9623 @joewallwork
> Here is the pdf re-generated with the typos in the references corrected.

Thanks, I can see that those points have now been addressed.

> As for the two issues in `LightKrylov_examples`, I'll try to get them fixed by tomorrow:
>
>     * Regarding the build failure, we'll simply mention in the `gfortran 14.3` and up is needed for now.
>
>     * Regarding the build script itself: I'll list the deps in the `README.md` file and modify the shell script to query the user for the desired package manager (`conda`, `pip` or `spack` for `fpm`) and assume the others have already been installed (`gfortran`, `mpi` and `make`) for the sake of simplicity. Would that works for you [@joewallwork](https://github.com/joewallwork) ?

Yes, those solutions sound reasonable to me, thanks.
