### 2026-01-26 #9623 @loiseaujc
@danielskatz and @ewu63 : I'm working as we speak on the last two bullets of [LightKrylov/issues/238](https://github.com/nekStab/LightKrylov/issues/238). I'll update the paper right after meeting. As for the first bullet point related to benchmarking, I can cook up an example using `gmres` to solve `Ax = b ` with both `LightKrylov` and `PETSc` but that will have to wait for a couple of days as I don't have `PETSc` on my computer right now.
