### 2026-01-26 #9623 @joewallwork
> [@joewallwork](https://github.com/joewallwork) - please confirm when these are done and your recommendation to publish is complete

@danielskatz I can confirm that the issues I mentioned have now been resolved.
