### 2026-01-27 #9623 @ewu63
@loiseaujc I don't think the benchmarking point I brought up needs to be addressed, I'm happy with the reason you stated and I think the examples serve a good purpose. The other two points should just require minor rewording of the relevant sentences and hopefully will not take much time. I have now checked all the items on my checklist and just waiting on the two points above.
