### 2026-01-27 c665b65
Added stdlib newest DOI.
