+++ paper/paper.bib
@@ -185,3 +185,49 @@
                 languages;Syntactics;Arrays;Ecosystems;Parallel processing},
     doi = {10.1109/MCSE.2022.3159862},
 }
+
+ @software{stdlib:zenodo,
+  author       = {Federico Perini and
+                  Jeremie Vandenplas and
+                  José Alves and
+                  William B. Clodius and
+                  Jing and
+                  Milan Curcic and
+                  Milan Skocic and
+                  Sebastian Ehlert and
+                  Aman Godara and
+                  Ondřej Čertík and
+                  Jean-Christophe Loiseau and
+                  Suprit S Jahagirdar and
+                  chuckyvt and
+                  Evan Voyles and
+                  ZUO Zhihua and
+                  gareth-nx and
+                  Izaak "Zaak" Beekman and
+                  Gabriel Brown and
+                  Tomohiro Degawa and
+                  Nathaniel Shaffer and
+                  Harris Snyder and
+                  Ian Giestas Pauli and
+                  scivision and
+                  HugoMVale and
+                  Juan and
+                  Arjen Markus and
+                  Martin Diehl and
+                  Minh Dao and
+                  Ivan Pribec and
+                  Lewis McMillan},
+  title        = {fortran-lang/stdlib: v0.8.1},
+  month        = jan,
+  year         = 2026,
+  publisher    = {Zenodo},
+  version      = {v0.8.1},
+  doi          = {10.5281/zenodo.18375788},
+  url          = {https://doi.org/10.5281/zenodo.18375788},
+  swhid        = {swh:1:dir:ad462291f63717c75a40569a62105932d66d0148
+                   ;origin=https://doi.org/10.5281/zenodo.18346788;vi
+                   sit=swh:1:snp:277224627b2991be40e536f116570368da28
+                   e586;anchor=swh:1:rel:0970445c37ae0062978f9b9bcf29
+                   41e50a23881f;path=fortran-lang-stdlib-0ede301
+                  },
+}   
