### 2026-01-27 5ca3213
Rephrased the PETSc paragraph.
