### 2026-01-27 #9623 @loiseaujc
@ewu63 : Here is the latest version of the paper with with a more positive note about `PETSc` build systems and toned-down perspectives. I've also included a new reference for `stdlib` since we have just released `stdlib v0.8.1` with a Zenodo-generated DOI.
