### 2026-01-27 #9623 @alexfikl
> and binding to several different languages exist (including C, Fortran, Python, Rust or Julia)

@loiseaujc Small typo in the new version: should be "bindings for".
