### 2026-01-28 90fdd90
case fixes in paper.bib
