+++ paper/paper.bib
@@ -12,8 +12,7 @@
 }
 
 @article{frantz-2023,
-    title = {Krylov {{Methods}} for {{Large-Scale Dynamical Systems}}: {{
-             Application}} in {{Fluid Dynamics}}},
+    title = {Krylov Methods for Large-Scale Dynamical Systems: Application in Fluid Dynamics},
     shorttitle = {Krylov {{Methods}} for {{Large-Scale Dynamical Systems}}},
     author = {Frantz, R. A. S. and Loiseau, J.-Ch. and Robinet, J.-Ch.},
     year = {2023},
@@ -25,7 +24,7 @@
 }
 
 @book{saad-2003,
-    title = {Iterative {{Methods}} for {{Sparse Linear Systems}}},
+    title = {Iterative Methods for Sparse Linear Systems},
     author = {Saad, Yousef},
     year = {2003},
     month = jan,
@@ -39,7 +38,7 @@
 }
 
 @article{ipsen-1998,
-    title = {The {{Idea Behind Krylov Methods}}},
+    title = {The Idea Behind {K}rylov Methods},
     author = {Ipsen, Ilse C. F. and Meyer, Carl D.},
     year = {1998},
     month = dec,
@@ -79,7 +78,7 @@
               Mills and Lawrence Mitchell and Todd Munson and Jose~E. Roman and
               Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith and
               Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
-    title = {{PETS}c {W}eb page},
+    title = {{PETS}c Web page},
     url = {https://petsc.org/},
     howpublished = {\url{https://petsc.org/}},
     year = {2025},
@@ -176,7 +175,7 @@
               Richardson, Brad and Snyder, Harris and Urban, John and Vandenplas,
               Jérémie},
     journal = {Computing in Science & Engineering},
-    title = {The State of Fortran},
+    title = {The State of {F}ortran},
     year = {2022},
     volume = {24},
     number = {2},
