### 2026-01-28 #9623 @danielskatz
@loiseaujc - Please see https://github.com/nekStab/LightKrylov/pull/239 for some suggested changes based on my proofreading as we get close to acceptance and publication
