### 2026-01-29 #9623 @loiseaujc
@danielskatz : I'll proceed with the Zenodo archival and subsequent steps this morning and will let you know as soon as I'm done.
