### 2026-01-29 9191fd2
Added Poisson image.
