### 2026-01-29 #9623 @loiseaujc
- Tagged release : [`v0.1.1`](https://github.com/nekStab/LightKrylov/releases/tag/v0.1.1)
- Zenodo archive : [link](https://zenodo.org/records/18412137) with doi:[10.5281/zenodo.18412137](https://doi.org/10.5281/zenodo.18412137)

I made sure the title, authors, ORCID etc are actually matching. Let me know if you need anything else.
