### 2026-01-29 b2518b1
Added zenodo doi badge.
