### 2026-01-29 #9623 @danielskatz
Congratulations to @loiseaujc (Jean-Christophe Loiseau) and co-authors on your JOSS publication!!

And thanks to @joewallwork, @ewu63, and @alexfikl for reviewing!
JOSS completely depends on volunteers - we couldn't do this without you
